brought all changes up to more recent branch of X3D. Can't get tests to compile and have ignored i2dsim option

+++ src/module_param.f90
@@ -139,7 +139,7 @@ module variables
        real(mytype), dimension(nx,ny,nz) :: t,u,r
        real(mytype), dimension(ny,nz):: s
        real(mytype), dimension(nx):: ff,fs,fw
-       real(mytype) :: lind
+       integer :: lind
      END SUBROUTINE DERIVATIVE_X
      SUBROUTINE DERIVATIVE_Y(t,u,r,s,ff,fs,fw,pp,nx,ny,nz,npaire,lind)
        use decomp_2d_constants, only : mytype
@@ -147,7 +147,7 @@ module variables
        real(mytype), dimension(nx,ny,nz) :: t,u,r
        real(mytype), dimension(nx,nz):: s
        real(mytype), dimension(ny):: ff,fs,fw,pp
-       real(mytype) :: lind
+       integer :: lind
      END SUBROUTINE DERIVATIVE_Y
      SUBROUTINE DERIVATIVE_YY(t,u,r,s,ff,fs,fw,nx,ny,nz,npaire,lind)
        use decomp_2d_constants, only : mytype
@@ -155,7 +155,7 @@ module variables
        real(mytype), dimension(nx,ny,nz) :: t,u,r
        real(mytype), dimension(nx,nz):: s
        real(mytype), dimension(ny):: ff,fs,fw
-       real(mytype) :: lind
+       integer :: lind
      END SUBROUTINE DERIVATIVE_YY
      SUBROUTINE DERIVATIVE_Z(t,u,r,s,ff,fs,fw,nx,ny,nz,npaire,lind)
        use decomp_2d_constants, only : mytype
@@ -163,7 +163,7 @@ module variables
        real(mytype), dimension(nx,ny,nz) :: t,u,r
        real(mytype), dimension(nx,ny):: s
        real(mytype), dimension(nz):: ff,fs,fw
-       real(mytype) :: lind
+       integer :: lind
      END SUBROUTINE DERIVATIVE_Z
   END INTERFACE
 
@@ -206,7 +206,7 @@ module variables
        real(mytype), dimension(nx,ny,nz) :: t,u,r
        real(mytype), dimension(ny,nz):: s
        real(mytype), dimension(nx):: ff,fs,fw
-       real(mytype) :: lind
+       integer :: lind
      END SUBROUTINE FILTER_X
      SUBROUTINE FILTER_Y(t,u,r,s,ff,fs,fw,nx,ny,nz,npaire,lind)
        use decomp_2d_constants, only : mytype
@@ -214,7 +214,7 @@ module variables
        real(mytype), dimension(nx,ny,nz) :: t,u,r
        real(mytype), dimension(nx,nz):: s
        real(mytype), dimension(ny):: ff,fs,fw
-       real(mytype) :: lind
+       integer :: lind
      END SUBROUTINE FILTER_Y
      SUBROUTINE FILTER_Z(t,u,r,s,ff,fs,fw,nx,ny,nz,npaire,lind)
        use decomp_2d_constants, only : mytype
@@ -222,7 +222,7 @@ module variables
        real(mytype), dimension(nx,ny,nz) :: t,u,r
        real(mytype), dimension(nx,ny):: s
        real(mytype), dimension(nz):: ff,fs,fw
-       real(mytype) :: lind
+       integer :: lind
      END SUBROUTINE FILTER_Z
   END INTERFACE
